@@ -1,81 +1,54 @@
-import { Component } from "react";
-import PropTypes from "prop-types";
-
-import MarvelService from "../../services/MarvelService";
+import { useState, useEffect, useRef } from "react";
+import useMarvelService from "../../services/MarvelService";
 import Spinner from "../spinner/Spinner";
 import ErrorMessage from "../errorMessage/ErrorMessage";
 import Skeleton from "../skeleton/Skeleton";
 import "./charInfo.scss";
-import thor from "../../resources/img/thor.jpeg";
 
-class CharInfo extends Component {
-	state = {
-		char: null,
-		loading: false,
-		error: false,
-	};
+const CharInfo = (props) => {
+	const [char, setChar] = useState(null);
 
-	marvelService = new MarvelService();
+	const { loading, error, getCharacter, clearError } = useMarvelService();
+	const prevCharIdRef = useRef();
 
-	componentDidMount() {
-		this.updateChar();
-	}
+	useEffect(() => {
+		updateChar();
+		prevCharIdRef.current = props.charId;
+	}, []);
 
-	componentDidUpdate(prevProps) {
-		if (this.props.charId !== prevProps.charId) {
-			this.updateChar();
+	useEffect(() => {
+		if (prevCharIdRef.current !== props.charId) {
+			updateChar();
+			prevCharIdRef.current = props.charId;
 		}
-	}
+	}, [props.charId]);
 
-	updateChar = () => {
-		const { charId } = this.props;
-
-		if (!charId) {
+	const updateChar = () => {
+		if (!props.charId) {
 			return;
 		}
-
-		this.onCharLoading();
-		this.marvelService.getCharacter(charId).then(this.onCharLoaded).catch(this.onError);
+		clearError();
+		getCharacter(props.charId).then(onCharLoaded);
 	};
 
-	onCharLoaded = (char) => {
-		this.setState({
-			char,
-			loading: false,
-		});
+	const onCharLoaded = (char) => {
+		setChar(char);
 	};
 
-	onCharLoading = () => {
-		this.setState({
-			loading: true,
-		});
-	};
+	const skeleton = char || loading || error ? null : <Skeleton />;
+	const errorMessage = error ? <ErrorMessage /> : null;
+	const spinner = loading ? <Spinner /> : null;
+	const content = !(loading || error || !char) ? <View char={char} /> : null;
 
-	onError = () => {
-		this.setState({
-			loading: false,
-			error: true,
-		});
-	};
-
-	render() {
-		const { char, loading, error } = this.state;
-
-		const skeleton = char || loading || error ? null : <Skeleton />;
-		const errorMessage = error ? <ErrorMessage /> : null;
-		const spinner = loading ? <Spinner /> : null;
-		const content = !(loading || error || !char) ? <View char={char} /> : null;
-
-		return (
-			<div className="char__info">
-				{skeleton}
-				{errorMessage}
-				{spinner}
-				{content}
-			</div>
-		);
-	}
-}
+	return (
+		<div className="char__info">
+			{skeleton}
+			{errorMessage}
+			{spinner}
+			{content}
+		</div>
+	);
+};
 
 const View = ({ char }) => {
 	const { name, description, thumbnail, homepage, wiki, comics } = char;
