@@ -1,63 +1,40 @@
-import { Component } from "react";
+import { useState, useEffect, useRef } from "react";
 import Spinner from "../spinner/Spinner";
 import ErrorMessage from "../errorMessage/ErrorMessage";
-import MarvelService from "../../services/MarvelService";
+import useMarvelService from "../../services/MarvelService";
 import "./charList.scss";
 
-class CharList extends Component {
-	state = {
-		charList: [],
-		loading: true,
-		error: false,
-		newItemLoading: false,
-		offset: 1420,
-		charEnded: false,
-		activeChar: null,
+const CharList = (props) => {
+	const [charList, setCharList] = useState([]);
+	const [newItemLoading, setNewItemLoading] = useState(false);
+	const [offset, setOffset] = useState(1420);
+	const [charEnded, setCharEnded] = useState(false);
+	const [activeChar, setActiveChar] = useState(null);
+
+	const { loading, error, getAllCharacters } = useMarvelService();
+
+	useEffect(() => {
+		onRequest(offset, true);
+	}, []);
+
+	const onRequest = (offset, initial) => {
+		initial ? setNewItemLoading(false) : setNewItemLoading(true);
+		getAllCharacters(offset).then(onCharListLoaded);
 	};
 
-	marvelService = new MarvelService();
-
-	componentDidMount() {
-		this.onRequest();
-	}
-
-	onRequest = (offset) => {
-		this.onCharListLoading();
-		this.marvelService.getAllCharacters(offset).then(this.onCharListLoaded).catch(this.onError);
-	};
-
-	onCharListLoading = () => {
-		this.setState({
-			newItemLoading: true,
-		});
-	};
-
-	onCharListLoaded = (newCharList) => {
+	const onCharListLoaded = (newCharList) => {
 		let ended = false;
 
 		if (newCharList.length < 9) {
 			ended = true;
 		}
 
-		this.setState(({ offset, charList }) => ({
-			charList: [...charList, ...newCharList],
-			loading: false,
-			newItemLoading: false,
-			offset: offset + 9,
-			charEnded: ended,
-		}));
+		setCharList((charList) => [...charList, ...newCharList]);
+		setNewItemLoading((newItemLoading) => false);
+		setOffset((offset) => offset + 9);
+		setCharEnded((charEnded) => ended);
 	};
-
-	onError = () => {
-		this.setState({
-			error: true,
-			loading: false,
-		});
-	};
-
-	// Этот метод создан для оптимизации,
-	// чтобы не помещать такую конструкцию в метод render
-	renderItems(arr) {
+	function renderItems(arr) {
 		const items = arr.map((item) => {
 			let imgStyle = {
 				objectFit: "cover",
@@ -70,12 +47,12 @@ class CharList extends Component {
 
 			return (
 				<li
-					className={`char__item ${this.state.activeChar === item.id ? "char__item_selected" : ""}`}
+					className={`char__item ${activeChar === item.id ? "char__item_selected" : ""}`}
 					key={item.id}
 					tabIndex={0}
 					onClick={() => {
-						this.props.onCharSelected(item.id);
-						this.setState({ activeChar: item.id });
+						props.onCharSelected(item.id);
+						setActiveChar((activeChar) => item.id);
 					}}
 				>
 					<img src={item.thumbnail} alt={item.name} style={imgStyle} />
@@ -87,29 +64,26 @@ class CharList extends Component {
 		return <ul className="char__grid">{items}</ul>;
 	}
 
-	render() {
-		const { charList, loading, error, offset, newItemLoading, charEnded } = this.state;
-		const items = this.renderItems(charList);
-		const errorMessage = error ? <ErrorMessage /> : null;
-		const spinner = loading ? <Spinner /> : null;
-		const content = !(loading || error) ? items : null;
+	const items = renderItems(charList);
 
-		return (
-			<div className="char__list">
-				{errorMessage}
-				{spinner}
-				{content}
-				<button
-					className="button button__main button__long"
-					style={{ display: charEnded ? "none" : "block" }}
-					disabled={newItemLoading}
-					onClick={() => this.onRequest(offset)}
-				>
-					<div className="inner">load more</div>
-				</button>
-			</div>
-		);
-	}
-}
+	const errorMessage = error ? <ErrorMessage /> : null;
+	const spinner = loading && !newItemLoading ? <Spinner /> : null;
+
+	return (
+		<div className="char__list">
+			{errorMessage}
+			{spinner}
+			{items}
+			<button
+				className="button button__main button__long"
+				style={{ display: charEnded ? "none" : "block" }}
+				disabled={newItemLoading}
+				onClick={() => onRequest(offset)}
+			>
+				<div className="inner">load more</div>
+			</button>
+		</div>
+	);
+};
 
 export default CharList;
