@@ -1,81 +1,53 @@
-import { Component } from "react";
-import MarvelService from "../../services/MarvelService";
+import { useState, useEffect, useRef } from "react";
+import useMarvelService from "../../services/MarvelService";
 import Spinner from "../spinner/Spinner";
 import ErrorMessage from "../errorMessage/ErrorMessage";
-
 import mjolnir from "../../resources/img/mjolnir.png";
 import "./randomChar.scss";
 
-class RandomChar extends Component {
-	state = {
-		char: {},
-		loading: true,
-		error: false,
+const RandomChar = () => {
+	const [char, setChar] = useState({});
+
+	const { loading, error, getCharacter, clearError } = useMarvelService();
+
+	const onCharLoaded = (char) => {
+		setChar(char);
 	};
 
-	marvelService = new MarvelService();
+	useEffect(() => {
+		updateChar();
+	}, []);
 
-	onCharLoaded = (char) => {
-		this.setState({
-			char,
-			loading: false,
-		});
-	};
-
-	componentDidMount() {
-		this.updateChar();
-		// this.timerId = setInterval(this.updateChar, 3000);
-	}
-
-	componentWillUnmount() {
-		clearInterval(this.timerId);
-	}
-
-	onError = () => {
-		this.setState({
-			loading: false,
-			error: true,
-		});
-	};
-
-	updateChar = () => {
-		this.setState({
-			loading: true,
-			error: false,
-		});
-
+	const updateChar = () => {
+		clearError();
 		const id = Math.floor(Math.random() * (1011400 - 1011000) + 1011000);
-
-		this.marvelService.getCharacter(id).then(this.onCharLoaded).catch(this.onError);
+		getCharacter(id).then(onCharLoaded);
 	};
 
-	render() {
-		const { char, loading, error } = this.state;
-		const errorMessage = error ? <ErrorMessage /> : null;
-		const spinner = loading ? <Spinner /> : null;
-		const content = !(loading || error) ? <View char={char} /> : null;
+	const errorMessage = error ? <ErrorMessage /> : null;
+	const spinner = loading ? <Spinner /> : null;
+	const content = !(loading || error) ? <View char={char} /> : null;
 
-		return (
-			<div className="randomchar">
-				{errorMessage}
-				{spinner}
-				{content}
-				<div className="randomchar__static">
-					<p className="randomchar__title">
-						Random character for today!
-						<br />
-						Do you want to get to know him better?
-					</p>
-					<p className="randomchar__title">Or choose another one</p>
-					<button className="button button__main" onClick={() => this.updateChar()}>
-						<div className="inner">try it</div>
-					</button>
-					<img src={mjolnir} alt="mjolnir" className="randomchar__decoration" />
-				</div>
+	return (
+		<div className="randomchar">
+			{errorMessage}
+			{spinner}
+			{content}
+			<div className="randomchar__static">
+				<p className="randomchar__title">
+					Random character for today!
+					<br />
+					Do you want to get to know him better?
+				</p>
+				<p className="randomchar__title">Or choose another one</p>
+				<button className="button button__main" onClick={() => updateChar()}>
+					<div className="inner">try it</div>
+				</button>
+				<img src={mjolnir} alt="mjolnir" className="randomchar__decoration" />
 			</div>
-		);
-	}
-}
+		</div>
+	);
+};
 
 const View = ({ char }) => {
 	const { name, description, thumbnail, homepage, wiki } = char;
