@@ -1,9 +1,7 @@
-import img from './error.gif';
+import img from "./error.png";
 
 const ErrorMessage = () => {
-    return (
-        <img src={img} alt='Error' />
-    )
-}
+	return <img src={img} alt="Error" />;
+};
 
 export default ErrorMessage;
