@@ -1,29 +1,33 @@
-class MarvelService {
-	_apiBase = "https://gateway.marvel.com:443/v1/public/";
-	_apiKey = "apikey=4fab0452b786e6a6fd4ea0afed8ce0bb";
-	_baseOffset = 210;
+import { useHttp } from "../hooks/http.hook";
 
-	getResource = async (url) => {
-		let res = await fetch(url);
+const useMarvelService = () => {
+	const { loading, request, error, clearError } = useHttp();
 
-		if (!res.ok) {
-			throw new Error(`Could not fetch ${url}, status: ${res.status}`);
-		}
+	const _apiBase = "https://gateway.marvel.com:443/v1/public/";
+	const _apiKey = "apikey=4fab0452b786e6a6fd4ea0afed8ce0bb";
+	const _baseOffset = 210;
 
-		return await res.json();
+	const getAllCharacters = async (offset = _baseOffset) => {
+		const res = await request(`${_apiBase}characters?limit=9&offset=${offset}&${_apiKey}`);
+		return res.data.results.map(_transformCharacter);
 	};
 
-	getAllCharacters = async (offset = this._baseOffset) => {
-		const res = await this.getResource(`${this._apiBase}characters?limit=9&offset=${offset}&${this._apiKey}`);
-		return res.data.results.map(this._transformCharacter);
+	const getCharacter = async (id) => {
+		const res = await request(`${_apiBase}characters/${id}?${_apiKey}`);
+		return _transformCharacter(res.data.results[0]);
 	};
 
-	getCharacter = async (id) => {
-		const res = await this.getResource(`${this._apiBase}characters/${id}?${this._apiKey}`);
-		return this._transformCharacter(res.data.results[0]);
+	const getAllComics = async (offset = 0) => {
+		const res = await request(`${_apiBase}comics?orderBy=issueNumber&limit=8&offset=${offset}&${_apiKey}`);
+		return res.data.results.map(_transformComics);
 	};
 
-	_transformCharacter = (char) => {
+	const getComic = async (id) => {
+		const res = await request(`${_apiBase}comics/${id}?${_apiKey}`);
+		return _transformComics(res.data.results[0]);
+	};
+
+	const _transformCharacter = (char) => {
 		return {
 			id: char.id,
 			name: char.name,
@@ -34,6 +38,28 @@ class MarvelService {
 			comics: char.comics.items,
 		};
 	};
-}
 
-export default MarvelService;
+	const _transformComics = (comics) => {
+		return {
+			id: comics.id,
+			title: comics.title,
+			description: comics.description || "There is no description",
+			pageCount: comics.pageCount ? `${comics.pageCount} p.` : "No information about the number of pages",
+			thumbnail: comics.thumbnail.path + "." + comics.thumbnail.extension,
+			language: comics.textObjects[0]?.language || "en-us",
+			price: comics.prices[0].price ? `${comics.prices[0].price}$` : "not available",
+		};
+	};
+
+	return {
+		loading,
+		error,
+		clearError,
+		getAllCharacters,
+		getCharacter,
+		getAllComics,
+		getComic,
+	};
+};
+
+export default useMarvelService;
